Replaces the CParser & add ArrayOf type.
The addition of the ArrayOf type was needed for the newest
specification. This resulted in having to edit the CParser again. As the
parser was not based on a formal specification of C, this was not that
easy. Therefore instead of editing it, I decided that it was better to
rewrite it using a formal grammar. The result is hopefully a more stable
and maintainable parser.